https://github.com/caxanga334/rcbot2/commit/d13409e016c38c0ffa0ebffab9618c697c9d8c13

+++ utils/RCBot2_meta/bot_task.cpp
@@ -2312,7 +2312,7 @@ void CBotTF2FindPipeWaypoint :: execute (CBot *pBot,CBotSchedule *pSchedule)
 				}
 
 				// If this waypoint is NOT visible to target it is good
-				if ( !m_pTable->GetVisibilityFromTo(m_iTargetWaypoint,m_WaypointsJ[m_j]) )
+				if ( !m_pTable->GetVisibilityFromTo((int)m_iTargetWaypoint,(int)m_WaypointsJ[m_j]) )
 				{			
 					m_fNearesti = fidist;
 					m_fNearestj = fjdist;
